Fixed  by using numbers instead of floats on total calculation.
This is a well known problem that you can encounter when working with floats. Essipecially in interprated programing languages. JS Math Model is not the only model that you can have this problem. My prefered method would be never using floats or doubles not even in DB when dealing with currencies. But for a quick fix this would also serves the purpose. I don't see any where that we do any other calcualtions similar like we do in `/orders/:id/order-total`